requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
openai>=1.47.0
python-dotenv>=1.0.1
redis>=5.0.1